SBoM code changes 3/14/2020
Pulled the following Functions out of analyze_data.py and created separate routines for each function.
1.	bar_image function pulled and created routine bar_img.py
2.	stacked_bar_image function pulled and created routine sbar_img.py
3.	multi_date_number_servers function pulled and created routine mdate_num_svrs.py
The following Functions from analyze_data.py, as well as analyze_data.py itself were abandoned
1.	print_dict_struc
2.	server_count
3.	count_findings
4.	list_packages
5.	critical_servers
Also removed draw_cve.py routine and removed any import items not being used on all of the routines
As a result of removing the analyze_data.py routine, I needed to update the following
1.	queryn_cve_bin.py with
a.	“from sbar_img import stacked_bar_image”
2.	queryn.numsvrs.py with
a.	“from mdate_num_svrs import multi_date_number_servers”
b.	“from sbar_img import stacked_bar_image”
3.	queryn_multiver.py with
a.	“from bar_img import bar_image”
4.	queryn_sup_cve.py with
a.	“from bar_img import bar_image”
Added the following routines
1.	queryn_cve_pkg.py – gives perspective of how many CVEs per package
2.	queryn_cve_svr.py – gives perspective of how many CVEs per server
3.	queryn_cve_tot.py – gives perspective of how many CVEs total
4.	query1_svr_grp.py – gives a health report on each server in the entire group
