+++ Src/query1_filt_cve.py
@@ -1,4 +1,4 @@
-## Copyright (c) 2019 AT&T Intellectual Property. All rights reserved.
+## Copyright (c) 2020 AT&T Intellectual Property. All rights reserved.
 
 import sys
 from load_db import load_graph
@@ -7,7 +7,6 @@ from load_db import pkg_subset
 from load_db import svr_subset
 from sbom_helpers import get_gdbpath
 from sbom_helpers import validate_file_access
-from analyze_data import bar_image
 import random
 
 ## For given day graph, summarize state of non-suppressed, non-multiversion, cve
