+++ Src/query1_multiver.py
@@ -1,10 +1,9 @@
-## Copyright (c) 2019 AT&T Intellectual Property. All rights reserved.
+## Copyright (c) 2020 AT&T Intellectual Property. All rights reserved.
 
 import sys
 from load_db import load_graph
 from sbom_helpers import mypprint
 from load_db import svr_pkgs
-from load_db import intermediate
 from load_db import get_hostname
 from load_db import get_group
 from sbom_helpers import get_gdbpath
