+++ Src/query1_svr_list.py
@@ -1,10 +1,8 @@
-## Copyright (c) 2019 AT&T Intellectual Property. All rights reserved.
+## Copyright (c) 2020 AT&T Intellectual Property. All rights reserved.
 
 import sys
 from load_db import load_graph
-from load_db import intermediate
 from load_db import get_groups
-from load_db import get_group
 from load_db import server_by_group
 from sbom_helpers import get_gdbpath
 from sbom_helpers import mypprint
