+++ Src/queryn.numsvrs.py
@@ -1,19 +1,10 @@
-## Copyright (c) 2019 AT&T Intellectual Property. All rights reserved.
+## Copyright (c) 2020 AT&T Intellectual Property. All rights reserved.
 
 ## given a list of gdb's, return number of servers by group and total
 
 import sys
-import numpy
-from load_db import load_graph
-from load_db import get_groups
-from load_db import intermediate
-import matplotlib.pyplot as plt
-from sbom_helpers import get_gdbpath
-from sbom_helpers import get_datelist
-from sbom_helpers import get_anlpath
-from sbom_helpers import validate_file_access
-from analyze_data import multi_date_number_servers
-from analyze_data import stacked_bar_image
+from mdate_num_svrs import multi_date_number_servers
+from sbar_img import stacked_bar_image
 
 ## output will be a dictionary (date as key) of dictionaries
 ##    per-date dictionaries will have keys of groups and total
@@ -38,5 +29,5 @@ params['filename'] = f
 params['title'] = 'Servers by Date'
 params['ylabel'] = 'Number of Servers'
 params['xlabel'] = 'Dates'
-params['colors'] = ['blue', 'purple', 'green', 'orange']
+params['colors'] = ['lightsteelblue', 'plum', 'palegreen', 'khaki', 'cyan']
 stacked_bar_image(data, params)
